[Cinelog / Room] now extract data from old db directly, instead of business services

+++ app/src/main/java/com/ulicae/cinelog/room/dao/TagDao.java
@@ -1,7 +1,6 @@
 package com.ulicae.cinelog.room.dao;
 
 import androidx.room.Dao;
-import androidx.room.Delete;
 import androidx.room.Query;
 
 import com.ulicae.cinelog.room.entities.Tag;
@@ -42,6 +41,6 @@ public interface TagDao extends AsyncRoomDao<Tag> {
     Flowable<List<Tag>> findSerieTags();
 
     @Query("SELECT * FROM tag WHERE id = :id")
-    Flowable<Tag> find(long id);
+    Flowable<Tag> find(int id);
 
 }
